Refactor data import processing and clean redundant code
This commit refactors the data import processing in Import.tpl view and Data/Import.php trait. It involves simplifying loops, removing redundant checks, and streamlining the import processing. In consequence, the code is now cleaner, easier to read and to maintain. Moreover, the validation and expose checks have been removed leaving it to more appropriate layers of code logic.